SASS-7607: Upgrade to play 3.0
