+++ it/test/connectors/GetOtherEmploymentsIncomeConnectorISpec.scala
@@ -18,41 +18,49 @@ package connectors
 
 import com.github.tomakehurst.wiremock.http.HttpHeader
 import config.{AppConfig, BackendAppConfig}
-import helpers.WiremockSpec
+import connectors.GetOtherEmploymentsIncomeConnector
 import models._
 import org.scalatestplus.play.PlaySpec
 import play.api.Configuration
 import play.api.http.Status._
-import play.api.libs.json.{JsObject, Json}
+import play.api.libs.json.Json
+import test.helpers.WiremockSpec
 import uk.gov.hmrc.http.{HeaderCarrier, HeaderNames, HttpClient, SessionId}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import utils.TaxYearUtils.{convertSpecificTaxYear, specificTaxYear}
+import utils.TaxYearUtils.convertSpecificTaxYear
 
-class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
+class GetOtherEmploymentsIncomeConnectorISpec extends PlaySpec with WiremockSpec {
 
-  lazy val connector: GetInsurancePoliciesTYSConnector = app.injector.instanceOf[GetInsurancePoliciesTYSConnector]
+  lazy val connector: GetOtherEmploymentsIncomeConnector = app.injector.instanceOf[GetOtherEmploymentsIncomeConnector]
   implicit val hc: HeaderCarrier = HeaderCarrier()
   val nino = "nino"
-  val taxYearParameter = convertSpecificTaxYear(specificTaxYear)
-  val url = s"/income-tax/insurance-policies/income/$taxYearParameter/$nino"
+  val taxYear = 2024
+  val taxYearParameter = convertSpecificTaxYear(taxYear)
+  val url = s"/income-tax/income/other/employments/$taxYearParameter/$nino"
 
   lazy val httpClient: HttpClient = app.injector.instanceOf[HttpClient]
 
   def appConfig(ifHost: String): AppConfig = new BackendAppConfig(app.injector.instanceOf[Configuration], app.injector.instanceOf[ServicesConfig]) {
   }
 
-  val ifReturned: InsurancePoliciesModel = InsurancePoliciesModel(
-    submittedOn = "2020-01-04T05:01:01Z",
-    lifeInsurance = Some(Seq(LifeInsuranceModel(Some("RefNo13254687"), Some("Life"), 123.45, Some(true), Some(4), Some(3), Some(123.45)))),
-    capitalRedemption = Some(Seq(CapitalRedemptionModel(Some("RefNo13254687"), Some("Capital"), 123.45, Some(true), Some(3), Some(2), Some(0)))),
-    lifeAnnuity = Some(Seq(LifeAnnuityModel(Some("RefNo13254687"), Some("Life"), 0, Some(true), Some(2), Some(22), Some(123.45)))),
-    voidedIsa = Some(Seq(VoidedIsaModel(Some("RefNo13254687"), Some("isa"), 123.45, Some(123.45), Some(5), Some(6)))),
-    foreign = Some(Seq(ForeignModel(Some("RefNo13254687"), 123.45, Some(123.45), Some(3))))
+  val ifReturned: OtherEmploymentsIncomeModel = OtherEmploymentsIncomeModel(
+    submittedOn = "2019-08-24T14:15:22Z",
+    shareOption = Some(Seq(ShareOptionModel("string", Some("123/abc 001<Q>"), "EMI", "2019-08-24", "2019-08-24", Some(true),
+      99999999999.99, 0, Some("string"), 99999999999.99, 99999999999.99, 99999999999.99, 99999999999.99,
+      99999999999.99, 99999999999.99))),
+    sharesAwardedOrReceived = Some(Seq(SharesAwardedOrReceivedModel("string", Some("123/abc 001<Q>"), "SIP", "2019-08-24", 0, "string",
+      "2019-08-24", true, true, 99999999999.99, 99999999999.99, 99999999999.99,
+      99999999999.99, 99999999999.99))),
+    disability = Some(DisabilityModel(Some("string"), 99999999999.99)),
+    foreignService = Some(ForeignServiceModel(Some("string"), 99999999999.99)),
+    lumpSums = Some(Seq(LumpSumsModel("string", "123/abc 001<Q>",
+      Some(TaxableLumpSumsAndCertainIncomeModel(99999999999.99, Some(99999999999.99), Some(true))),
+      Some(BenefitFromEmployerFinancedRetirementSchemeModel(99999999999.99, Some(99999999999.99), Some(99999999999.99), Some(true))),
+      Some(RedundancyCompensationPaymentsOverExemptionModel(99999999999.99, Some(99999999999.99), Some(true))),
+      Some(RedundancyCompensationPaymentsUnderExemptionModel(99999999999.99)))))
   )
 
-  val ifReturnedEmpty: JsObject = Json.obj()
-
-  " GetInsurancePoliciesTYSConnector" should {
+  " GetOtherEmploymentsIncomeConnector" should {
 
     "include internal headers" when {
 
@@ -68,7 +76,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
 
         stubGetWithResponseBody(url, OK, Json.toJson(ifReturned).toString)
 
-        val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+        val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
         result mustBe Right(ifReturned)
       }
@@ -78,9 +86,9 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
 
         stubGetWithResponseBody(url, OK, Json.toJson(ifReturned).toString, headersSentToDes)
 
-        val connector = new GetInsurancePoliciesTYSConnector(httpClient, appConfig(externalHost))
+        val connector = new GetOtherEmploymentsIncomeConnector(httpClient, appConfig(externalHost))
 
-        val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+        val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
         result mustBe Right(ifReturned)
       }
@@ -90,21 +98,9 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
 
       "IF returns a 200" in {
         stubGetWithResponseBody(url, OK, Json.toJson(ifReturned).toString)
-        val result = await(connector.getInsurancePolicies(nino, specificTaxYear))
+        val result = await(connector.getOtherEmploymentsIncome(nino, taxYear))
 
         result mustBe Right(ifReturned)
-
-      }
-    }
-
-    "return an error" when {
-
-      "IF returns an empty 200" in {
-        stubGetWithResponseBody(url, OK, ifReturnedEmpty.toString())
-        val result = await(connector.getInsurancePolicies(nino, specificTaxYear))
-        val expectedResult = ErrorModel(INTERNAL_SERVER_ERROR, ErrorBodyModel.parsingError)
-
-        result mustBe Left(expectedResult)
 
       }
     }
@@ -115,7 +111,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, NO_CONTENT, "{}")
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -130,7 +126,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, BAD_REQUEST, responseBody.toString())
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -145,7 +141,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, UNPROCESSABLE_ENTITY, responseBody.toString())
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -160,7 +156,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, NOT_FOUND, responseBody.toString())
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -175,7 +171,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, INTERNAL_SERVER_ERROR, responseBody.toString())
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
@@ -190,7 +186,7 @@ class GetInsurancePoliciesTYSConnectorISpec extends PlaySpec with WiremockSpec {
       stubGetWithResponseBody(url, SERVICE_UNAVAILABLE, responseBody.toString())
 
       implicit val hc: HeaderCarrier = HeaderCarrier()
-      val result = await(connector.getInsurancePolicies(nino, specificTaxYear)(hc))
+      val result = await(connector.getOtherEmploymentsIncome(nino, taxYear)(hc))
 
       result mustBe Left(expectedResult)
     }
